Create shared memory with random key. Automatically removed.
If the default key is used and no destruction is forced the shared
memory persists across test-runs, and even across applications if you
aren't careful. For these tests that need to count accurately number
of requests this is somewhat of an issue. Creating a segment with a
random key and destroying it after the tests terminates circumvent
this problem.